+++ packages/orq-rc/src/orq_ai_sdk/agents.py
@@ -3,6 +3,11 @@
 from .basesdk import BaseSDK
 from orq_ai_sdk import models, utils
 from orq_ai_sdk._hooks import HookContext
+from orq_ai_sdk.models import (
+    listagenttasksop as models_listagenttasksop,
+    runagentop as models_runagentop,
+    streamrunagentop as models_streamrunagentop,
+)
 from orq_ai_sdk.types import BaseModel, OptionalNullable, UNSET
 from orq_ai_sdk.utils import eventstreaming, get_security_from_env
 from orq_ai_sdk.utils.unmarshal_json_response import unmarshal_json_response
@@ -584,7 +589,7 @@ class Agents(BaseSDK):
         self,
         *,
         agent_key: str,
-        status: List[models.Status],
+        status: List[models_listagenttasksop.Status],
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
@@ -699,7 +704,7 @@ class Agents(BaseSDK):
         self,
         *,
         agent_key: str,
-        status: List[models.Status],
+        status: List[models_listagenttasksop.Status],
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
@@ -814,7 +819,10 @@ class Agents(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[models.RunAgentRequestBody, models.RunAgentRequestBodyTypedDict]
+            Union[
+                models_runagentop.RunAgentRequestBody,
+                models_runagentop.RunAgentRequestBodyTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -907,7 +915,10 @@ class Agents(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[models.RunAgentRequestBody, models.RunAgentRequestBodyTypedDict]
+            Union[
+                models_runagentop.RunAgentRequestBody,
+                models_runagentop.RunAgentRequestBodyTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -1001,8 +1012,8 @@ class Agents(BaseSDK):
         *,
         request: Optional[
             Union[
-                models.StreamRunAgentRequestBody,
-                models.StreamRunAgentRequestBodyTypedDict,
+                models_streamrunagentop.StreamRunAgentRequestBody,
+                models_streamrunagentop.StreamRunAgentRequestBodyTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -1114,8 +1125,8 @@ class Agents(BaseSDK):
         *,
         request: Optional[
             Union[
-                models.StreamRunAgentRequestBody,
-                models.StreamRunAgentRequestBodyTypedDict,
+                models_streamrunagentop.StreamRunAgentRequestBody,
+                models_streamrunagentop.StreamRunAgentRequestBodyTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,