+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -140,7 +140,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-02T13:55:24.060Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-02T14:54:45.724Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")