+++ packages/orq-rc/RELEASES.md
@@ -3089,3 +3089,13 @@ Based on:
 - [python v3.14.0-rc.3] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.3] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.3 - packages/orq-rc
+
+## 2025-10-02 14:56:35
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.633.2 (2.720.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.4] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.4] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.4 - packages/orq-rc